refactor: 상품 정보를 커스텀 훅 형태로 분리
